fix warnings with Intel C++
git-svn-id: file:///home/git-svn-mirror/pushmi-svn-mirror/trunk@39438 dd0e9695-b195-4be7-bd10-2dea1a65a6b6

+++ tools/switchboard2/main.cc
@@ -149,7 +149,7 @@ main(int argc, char **argv)
 			}
 
 			pthread_t tid;
-			pthread_create(&tid, NULL, acceptor_thread, reinterpret_cast<void *>(lsnsock));
+			pthread_create(&tid, NULL, acceptor_thread, reinterpret_cast<void *>((intptr_t) lsnsock));
 		} else {
 			int r;
 			struct addrinfo hints, *res, *iter;
@@ -202,7 +202,7 @@ main(int argc, char **argv)
 				}
 
 				pthread_t tid;
-				pthread_create(&tid, NULL, acceptor_thread, reinterpret_cast<void *>(lsnsock));
+				pthread_create(&tid, NULL, acceptor_thread, reinterpret_cast<void *>((intptr_t) lsnsock));
 			}
 
 			freeaddrinfo(res);
